Add HttpClient to fetch and display medication list
The MedListController has been updated to use HttpClient for making a GET request to an external API. The fetched JSON data is deserialized and passed into the view for display. This data is a list of medication names, which are properly shown in the updated Index view. The HttpClient service is registered in the SleepEditWeb.Program file.